# openapi.json
{
  "components" : {
    "schemas" : {
...
      "ApiResponseExtractWordsResponse" : {
        "properties" : {
          "code" : {
            "type" : [ "string", "null" ]
          },
          "data" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/ExtractWordsResponse"
            }, {
              "type" : "null"
            } ]
          },
          "message" : {
            "type" : [ "string", "null" ]
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "required" : [ "success" ],
        "type" : "object"
      },
...
      "ApiResponseSuggestVocabularyResponse" : {
        "properties" : {
          "code" : {
            "type" : [ "string", "null" ]
          },
          "data" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/SuggestVocabularyResponse"
            }, {
              "type" : "null"
            } ]
          },
          "message" : {
            "type" : [ "string", "null" ]
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "required" : [ "success" ],
        "type" : "object"
      },
...
      "ApiResponseTranslateTextResponse" : {
        "properties" : {
          "code" : {
            "type" : [ "string", "null" ]
          },
          "data" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/TranslateTextResponse"
            }, {
              "type" : "null"
            } ]
          },
          "message" : {
            "type" : [ "string", "null" ]
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "required" : [ "success" ],
        "type" : "object"
      },
...
      "ApiResponseVocabularyExtractionResponse" : {
        "properties" : {
          "code" : {
            "type" : [ "string", "null" ]
          },
          "data" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/VocabularyExtractionResponse"
            }, {
              "type" : "null"
            } ]
          },
          "message" : {
            "type" : [ "string", "null" ]
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "required" : [ "success" ],
        "type" : "object"
      },
...
      "ExtractVocabularyRequest" : {
        "properties" : {
          "extractSentences" : {
            "type" : "boolean"
          },
          "extractWords" : {
            "type" : "boolean"
          },
          "imageBase64" : {
            "minLength" : 1,
            "type" : "string"
          },
          "targetLanguage" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "extractSentences", "extractWords", "imageBase64", "targetLanguage" ],
        "type" : "object"
      },
      "ExtractWordsRequest" : {
        "properties" : {
          "imageBase64" : {
            "minLength" : 1,
            "type" : "string"
          },
          "includePhrases" : {
            "type" : "boolean"
          },
          "learningLanguage" : {
            "minLength" : 1,
            "type" : "string"
          },
          "nativeLanguage" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "imageBase64", "includePhrases", "learningLanguage", "nativeLanguage" ],
        "type" : "object"
      },
      "ExtractWordsResponse" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/ExtractedWordItem"
            },
            "type" : "array"
          }
        },
        "required" : [ "items" ],
        "type" : "object"
      },
      "ExtractedWordItem" : {
        "properties" : {
          "note" : {
            "type" : "string"
          },
          "term" : {
            "type" : "string"
          },
          "translation" : {
            "type" : "string"
          }
        },
        "required" : [ "note", "term", "translation" ],
        "type" : "object"
      },
...
      "SuggestVocabularyItemResponse" : {
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "originalWord" : {
            "type" : "string"
          },
          "translation" : {
            "type" : "string"
          }
        },
        "required" : [ "description", "originalWord", "translation" ],
        "type" : "object"
      },
      "SuggestVocabularyRequest" : {
        "properties" : {
          "currentLevel" : {
            "minLength" : 1,
            "type" : "string"
          },
          "interests" : {
            "items" : {
              "type" : "string"
            },
            "maxItems" : 10,
            "minItems" : 0,
            "type" : "array"
          },
          "nativeLanguage" : {
            "minLength" : 1,
            "type" : "string"
          },
          "targetLanguage" : {
            "minLength" : 1,
            "type" : "string"
          },
          "targetLanguageCode" : {
            "maxLength" : 8,
            "minLength" : 0,
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "currentLevel", "interests", "nativeLanguage", "targetLanguage" ],
        "type" : "object"
      },
      "SuggestVocabularyResponse" : {
        "properties" : {
          "currentLevel" : {
            "type" : "string"
          },
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/SuggestVocabularyItemResponse"
            },
            "type" : "array"
          },
          "nativeLanguage" : {
            "type" : "string"
          },
          "targetLanguage" : {
            "type" : "string"
          }
        },
        "required" : [ "currentLevel", "items", "nativeLanguage", "targetLanguage" ],
        "type" : "object"
      },
...
      "TranslateTextRequest" : {
        "properties" : {
          "targetLanguage" : {
            "minLength" : 1,
            "type" : "string"
          },
          "text" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "targetLanguage", "text" ],
        "type" : "object"
      },
      "TranslateTextResponse" : {
        "properties" : {
          "originalText" : {
            "type" : "string"
          },
          "translation" : {
            "type" : "string"
          }
        },
        "required" : [ "originalText", "translation" ],
        "type" : "object"
      },
...
      "VocabularyExtractionResponse" : {
        "properties" : {
          "extractedText" : {
            "type" : "string"
          },
          "wordCount" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "extractedText", "wordCount" ],
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/ai/extract-vocabulary" : {
      "post" : {
        "operationId" : "extractVocabulary",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ExtractVocabularyRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVocabularyExtractionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "ai-controller" ]
      }
    },
    "/api/v1/ai/extract-words" : {
      "post" : {
        "operationId" : "extractWords",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ExtractWordsRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseExtractWordsResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "ai-controller" ]
      }
    },
...
    "/api/v1/ai/suggest-vocabulary" : {
      "post" : {
        "operationId" : "suggestVocabulary",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SuggestVocabularyRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseSuggestVocabularyResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "ai-controller" ]
      }
    },
    "/api/v1/ai/translate-text" : {
      "post" : {
        "operationId" : "translateText",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TranslateTextRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTranslateTextResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "ai-controller" ]
      }
    },
...
  }
}